refactor: add path alias to src & move PrismaService to separate module - Add `@/` as path alias to `src` folder - Create `PrismaModule` that export the `PrismaService`
