+++ src/island/island.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { IslandController } from './island.controller';
 import { IslandService } from './island.service';
-import { PrismaService } from '../common/services/prisma';
+import { PrismaService } from '@/prisma/prisma.service';
 
 describe('IslandController', () => {
   let controller: IslandController;
