+++ src/island/island.module.ts
@@ -1,10 +1,11 @@
+import { PrismaModule } from '@/prisma/prisma.module';
 import { Module } from '@nestjs/common';
 import { IslandController } from './island.controller';
 import { IslandService } from './island.service';
-import { PrismaService } from '../common/services/prisma';
 
 @Module({
-  providers: [IslandService, PrismaService],
+  imports: [PrismaModule],
+  providers: [IslandService],
   controllers: [IslandController],
   exports: [IslandService],
 })
