+++ src/regency/regency.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from '~/src/common/services/prisma';
 import { IslandModule } from '../island/island.module';
 import { RegencyController } from './regency.controller';
 import { RegencyService } from './regency.service';
+import { PrismaModule } from '@/prisma/prisma.module';
 
 @Module({
-  imports: [IslandModule],
+  imports: [PrismaModule, IslandModule],
   controllers: [RegencyController],
-  providers: [PrismaService, RegencyService],
+  providers: [RegencyService],
 })
 export class RegencyModule {}
