+++ src/village/village.dto.ts
@@ -1,6 +1,6 @@
 import { IsNotEmpty, IsNumberString, Length } from 'class-validator';
-import { EqualsAny } from '~/src/common/decorator/EqualsAny';
-import { IsNotSymbol } from '~/src/common/decorator/IsNotSymbol';
+import { EqualsAny } from '@/common/decorator/EqualsAny';
+import { IsNotSymbol } from '@/common/decorator/IsNotSymbol';
 import { SortQuery } from '~/utils/helpers/sorter';
 import { IntersectionType, PickType } from '@nestjs/mapped-types';
