+++ src/village/village.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { Village } from '@prisma/client';
 import { Sorter, SortOptions } from '~/utils/helpers/sorter';
-import { PrismaService } from '~/src/common/services/prisma';
+import { PrismaService } from '@/prisma/prisma.service';
 import { getDBProviderFeatures } from '~/utils/db';
 
 type VillageSortKeys = keyof Village;
